requirements.txt
automata==0.1.4
automata_lib==5.0.0
jsonpickle==2.0.0
numpy==1.21.3
opencv_python==4.5.4.58
Pillow==8.4.0
pydot==1.4.2
streamlit==1.1.0
visual_automata==1.1.1
